agents/Day-06-Learning-Coach/backend/app/core/agent.py
"""
Core agent implementation for the Learning Coach.
"""

from typing import Dict, List, Any, Optional, TypedDict, Union
from typing_extensions import Annotated

from langchain.chains import LLMChain
from langchain_core.prompts import PromptTemplate
from langchain_core.runnables import RunnablePassthrough
from langgraph.graph import StateGraph, END


# Define the state schema as a TypedDict
class AgentState(TypedDict):
    """State for the Learning Coach Agent."""
    user_input: str
    intent: Optional[str]
    context: Dict[str, Any]
    response: Optional[str]


class LearningCoachAgent:
    """Learning Coach Agent implementation using LangChain and LangGraph."""

    def __init__(self):
        """Initialize the Learning Coach Agent."""
        self.state_graph = self._build_state_graph()

    def _build_state_graph(self) -> StateGraph:
        """Build the state graph for the agent workflow.

        Returns:
            StateGraph: The state graph for the agent workflow.
        """
        # Create the state graph with the TypedDict schema
        graph = StateGraph(AgentState)

        # Define a simple placeholder node for now
        def placeholder_node(state: AgentState) -> AgentState:
            """Simple placeholder node that just echoes the input."""
            state["response"] = f"I received your input: {state['user_input']}. This is a placeholder response."
            return state

        # Add the node to the graph
        graph.add_node("placeholder", placeholder_node)

        # Define the entry point
        graph.set_entry_point("placeholder")

        # Define the exit point
        graph.add_edge("placeholder", END)

        # Compile the graph
        return graph.compile()

    async def process_input(self, user_input: str, context: Dict[str, Any] = None) -> Dict[str, Any]:
        """Process user input and generate a response.

        Args:
            user_input: The user's input text
            context: Optional context information

        Returns:
            Dict containing the agent's response and any additional information
        """
        # Initialize context if not provided
        if context is None:
            context = {}

        # Create initial state
        initial_state: AgentState = {
            "user_input": user_input,
            "intent": None,
            "context": context,
            "response": None,
        }

        # Run the state graph with the initial state
        final_state = self.state_graph.invoke(initial_state)

        # Return the response and context
        return {
            "response": final_state["response"] or "I'm not sure how to respond to that.",
            "context": final_state["context"],
        }
