agents/Day-06-Learning-Coach/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langgraph>=0.4.3",
    "pydantic[email]>=2.11.4",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "typing-extensions>=4.13.2",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
]
